fix reflection bug
